+++ src/onegov.ballot/onegov/ballot/__init__.py
@@ -1,10 +1,24 @@
-from onegov.ballot.models import Ballot, BallotResult, Vote
+from onegov.ballot.models import (
+    Ballot,
+    BallotResult,
+    CandidateResult,
+    Election,
+    ElectionResult,
+    ListResult,
+    Vote
+)
+from onegov.ballot.collection import ElectionCollection
 from onegov.ballot.collection import BallotCollection, VoteCollection
 
 __all__ = [
     'Ballot',
     'BallotCollection',
     'BallotResult',
+    'CandidateResult',
+    'Election',
+    'ElectionCollection',
+    'ElectionResult',
+    'ListResult',
     'Vote',
     'VoteCollection'
 ]