+++ src/onegov.ballot/onegov/ballot/collection.py
@@ -1,5 +1,73 @@
-from onegov.ballot.models import Ballot, Vote
+from onegov.ballot.models import Ballot, Election, Vote
 from sqlalchemy import cast, desc, extract, Integer
+
+
+class ElectionCollection(object):
+
+    def __init__(self, session, year=None):
+        self.session = session
+        self.year = year
+
+    def query(self):
+        return self.session.query(Election)
+
+    def for_year(self, year):
+        return self.__class__(self.session, year)
+
+    def get_latest(self):
+        """ Returns the elections with the latest date. """
+
+        latest_date = self.query().with_entities(Election.date)
+        latest_date = latest_date.order_by(desc(Election.date))
+        latest_date = latest_date.limit(1).first()
+
+        if not latest_date:
+            return None
+        else:
+            return self.by_date(latest_date)
+
+    def get_years(self):
+        """ Returns a list of years for which there are elections. """
+
+        query = self.query()
+        query = query.with_entities(
+            cast(extract('year', Election.date), Integer)
+        )
+        query = query.order_by(desc(Election.date))
+
+        return list(reversed(sorted((set(r[0] for r in query.all())))))
+
+    def by_date(self, date):
+        """ Returns the elections on the given date. """
+
+        query = self.query()
+        query = query.filter(Election.date == date)
+        query = query.order_by(
+            Election.domain, Election.shortcode, Election.title
+        )
+
+        return query.all()
+
+    def by_year(self, year=None):
+        """ Returns the elections for the current/given year. """
+
+        year = year or self.year
+
+        query = self.query()
+        query = query.filter(extract('year', Election.date) == year)
+        query = query.order_by(
+            Election.date, Election.domain, Election.shortcode, Election.title
+        )
+
+        return query.all()
+
+    def by_id(self, id):
+        """ Returns the election by id. """
+
+        query = self.query()
+        query = query.filter(Election.id == id)
+
+        return query.first()
 
 
 class VoteCollection(object):